+++ indie2dEngine-Core/Sources/CSceneFabricator.h
@@ -39,7 +39,7 @@ public:
     CSceneFabricator(std::shared_ptr<CTemplateAccessor> _templateAccessor, std::shared_ptr<CResourceAccessor> _resourceAccessor);
     virtual ~CSceneFabricator(void);
     
-    std::shared_ptr<CCamera> CreateCamera(f32 _fov, f32 _near, f32 _far,const glm::vec4& _viewport);
+    std::shared_ptr<CCamera> CreateCamera(f32 _fov, f32 _near, f32 _far,const glm::ivec4& _viewport);
     void DeleteCamera(std::shared_ptr<CCamera> _camera);
     
     std::shared_ptr<CLight> CreateLight(void);
